// RijksApp/src/app/interfaces/selected-filter.ts
export interface SelectedFilter {
    involvedMaker: string | undefined
    material: string | undefined
    type: string | undefined
    technique: string | undefined
    datingPeriod: number | undefined
    normalized32ColorHex: string | undefined
    imgonly: boolean | undefined
    toppieces: boolean | undefined,
    sort: string | undefined
}